refactor: 초기 data와 type 분리
